==> s2_feat_engineering.ipynb <==
# Бинарная классификация мошеннических транзакций
## Часть 2. Инжиниринг данных и отбор признаков

### План

1. [Исходные данные]
2. [Разделение на train/test]
3. [Feature Engineering на train]
4. [Отбор признаков на train]
5. [Ребалансировка **только** train]
6. [Обучение модели]
7. [Оценка на **оригинальном** test]






import pandas as pd
import numpy as np
from sklearn.cluster import KMeans

df_train_scaled = pd.read_csv("train_scaled.csv")
df_original_test = pd.read_csv("test_scaled.csv")

### Train Test Split

X = df_train_scaled.drop(columns=['Class'])
y = df_train_scaled['Class']

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    random_state=42,
    stratify=y  # стратификация!
)

### Инжиниринг новых признаков на тренировочных данных

# Взаимодействия и полиномы
X_train['V6_V27'] = X_train['V6'] * X_train['V27']
X_train['Amount_sq'] = X_train['Amount'] ** 2

# Бинарные маркеры выбросов (3σ)
for col in ['V6', 'Amount']:
    X_train[f'IsOutlier_{col}'] = (np.abs(X_train[col]) > 3).astype(int)

# Добавим
train_std = {'V6': X_train['V6'].std(), 'Amount': X_train['Amount'].std()}

# Логарифмирование суммы
X_train['Log_Amount'] = np.log1p(X_train['Amount'])

# Кластеризация K-means
kmeans = KMeans(n_clusters=5, random_state=42)
kmeans.fit(X_train[['V6', 'V20', 'V23', 'V27']])
X_train['Cluster'] = kmeans.predict(X_train[['V6', 'V20', 'V23', 'V27']])

### Инжиниринг новых признаков на валидационных данных

X_test['V6_V27'] = X_test['V6'] * X_test['V27']
X_test['Amount_sq'] = X_test['Amount'] ** 2

# Бинарные маркеры выбросов. Обратить внимание на порог `> 3 * train_std[col]`
for col in ['V6', 'Amount']:
    X_test[f'IsOutlier_{col}'] = (np.abs(X_test[col]) > 3 * train_std[col]).astype(int)

# Логарифмирование суммы
X_test['Log_Amount'] = np.log1p(X_test['Amount'])

# Кластеризация (используем обученный кластеризатор!)
X_test['Cluster'] = kmeans.predict(X_test[['V6', 'V20', 'V23', 'V27']])

### Инжиниринг новых признаков на тестовых данных

df_original_test['V6_V27'] = df_original_test['V6'] * df_original_test['V27']
df_original_test['Amount_sq'] = df_original_test['Amount'] ** 2

for col in ['V6', 'Amount']:
    df_original_test[f'IsOutlier_{col}'] = (np.abs(df_original_test[col]) > 3 * train_std[col]).astype(int)

df_original_test['Log_Amount'] = np.log1p(df_original_test['Amount'])
df_original_test['Cluster'] = kmeans.predict(df_original_test[['V6', 'V20', 'V23', 'V27']])

**Итог**: взаимодействия (`V6_V27`), полиномы (`Amount_sq`), бинарные маркеры выбросов (`IsOutlier_V6`), логарифмированная сумма (`Log_Amount`) и кластеры (`Cluster`) — созданы для выявления нелинейных зависимостей и аномалий.

### Отбор признаков

from lightgbm import LGBMClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.feature_selection import SelectFromModel
from sklearn.model_selection import train_test_split

# LightGBM Feature Importance
lgb = LGBMClassifier(random_state=42, class_weight='balanced')
lgb.fit(X_train, y_train)
importance = pd.Series(lgb.feature_importances_, index=X_train.columns)
selected_features_imp = importance.nlargest(15).index.tolist()

# L1-регуляризация
# selector = SelectFromModel(
#     estimator=LogisticRegression(
#         penalty='l1',
#         solver='saga',
#         class_weight='balanced',
#         max_iter=500,
#         random_state=42
#     ),
#     threshold="median"
# )



# selector.fit(X_train, y_train)
# selected_features_l1 = X.columns[selector.get_support()]

# print("Выбрано признаков:", sum(selector.get_support()))

**Методы отбора**:
*Feature Importance* (LightGBM) и [~L1-регуляризация~] — выбрали признаки с наибольшим влиянием на целевую переменную и устранили мультиколлинеарность.

# Итоговый выбор / объединение признаков
final_features = list(
    set(selected_features_imp) |
    # set(selected_features_l1) |
    {'V6_V27', 'Cluster', 'IsOutlier_V6'}
)
final_features

**В итоговый датасет вошли**: `V6_V27`, `Cluster`, `IsOutlier_V6`, а также ключевые исходные признаки ( `V3`, `V14` и др.), как наиболее значимые для разделения классов

import json

missing_in_train = set(final_features) - set(X_train.columns)
missing_in_test = set(final_features) - set(X_test.columns)

if missing_in_train:
    raise ValueError(f"Признаки отсутствуют в трейне: {missing_in_train}")
if missing_in_test:
    raise ValueError(f"Признаки отсутствуют в тесте: {missing_in_test}")

X_train_final = X_train[final_features]
X_test_final = X_test[final_features]

# Внесение изменений в оригинальный test dataset
df_original_test = df_original_test[final_features]
df_original_test.to_csv("test_feat_select.csv", index=False)

# Сохранение датасета с отобранными признаками
train_df = pd.concat([X_train_final, y_train], axis=1)
test_df = pd.concat([X_test_final, y_test], axis=1)
final_df = pd.concat([train_df, test_df], axis=0)
final_df.to_csv("train_feat_select.csv", index=False)

# Сохранение списка признаков
with open('selected_features.json', 'w') as f:
    json.dump(list(final_features), f)

del X_train_final, X_test_final, df_train_scaled, df_test_scaled, train_df, test_df, final_df

## Логистическая регрессия

import pandas as pd

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import accuracy_score, f1_score
from sklearn.model_selection import train_test_split
from imblearn.pipeline import Pipeline
from sklearn.metrics import classification_report
from imblearn.over_sampling import SMOTE

df = pd.read_csv("train_feat_select.csv")
X = df.drop(columns=['Class'])
y = df['Class']

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    random_state=42,
    stratify=y  # стратификация для баланса классов
)

### Ребалансировка классов

from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline

over = SMOTE(sampling_strategy=0.1, random_state=42)  # Доводим до 10% меньшего класса
under = RandomUnderSampler(sampling_strategy=0.5)     # Затем уменьшаем больший класс до соотношения 1:2

pipeline = Pipeline([
    ('o', over),
    ('u', under)
])

X_train_resampled, y_train_resampled = pipeline.fit_resample(X_train, y_train)

lr_model = LogisticRegression(
    solver="saga",
    class_weight="balanced",
    tol=1e-2,
    max_iter=500,
    random_state=42
)

param_grid = {
    "penalty": ["l1", "l2"],
    "C": [0.01, 0.1, 1.0],  # Сила регуляризации
    "solver": ["liblinear", "saga"]  # Для L1/L2
}

grid_search = GridSearchCV(
    estimator=lr_model,
    param_grid=param_grid,
    cv=5,
    scoring="recall",
    n_jobs=-1
)

# Обучение на тренировочных данных
grid_search.fit(X_train_resampled, y_train_resampled)

print("Лучшие параметры:", grid_search.best_params_)
print("Лучшая recall:", grid_search.best_score_)

### Отчёт

from sklearn.metrics import classification_report

y_pred = grid_search.predict(X_test)
print(classification_report(y_test, y_pred))

### Итог:

Произведены следующие этапы:

1. Разделение на train/test
2. Feature Engineering
3. Отбор признаков с помощью LightGBM
4. Ребалансировка классов в связи с сильным диcбалансом
5. Обучение модели логистической регрессии с подбором гиперпараметров через GridSearch, используя scoring="recall"
