==> s4_praktikum_update.ipynb <==
import pandas as pd

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import accuracy_score, f1_score
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.model_selection import GridSearchCV, StratifiedKFold
from sklearn.metrics import make_scorer, f1_score, precision_score, recall_score, roc_auc_score

df = pd.read_csv("train_with_dbscan_outliers.csv")
X = df.drop(columns=['Class'])
y = df['Class']

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    random_state=42,
    stratify=y  # стратификация для баланса классов
)

rf = RandomForestClassifier(
    random_state=42,
    class_weight='balanced',
    n_jobs=-1
)

param_grid = {
    'n_estimators': [10, 50, 100],
    'max_depth': [5, 10, 20, 30], 
}

scorers = {
    'f1': make_scorer(f1_score),
    'precision': make_scorer(precision_score),
    'recall': make_scorer(recall_score),
    'roc_auc': make_scorer(roc_auc_score)
}

cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

grid_search = GridSearchCV(
    estimator=rf,
    param_grid=param_grid,
    scoring='recall',
    cv=cv,
    n_jobs=-1,
    verbose=2
)

grid_search.fit(X_train, y_train)

grid_search.best_estimator_

y_pred = grid_search.predict(X_test)
y_proba = grid_search.predict_proba(X_test)[:, 1]

print(f"Подобранные параметры: {grid_search.best_params_}")
print(f"F1-score (CV): {grid_search.best_score_:.4f}")
print("\nМетрики:")
print(f"F1-score: {f1_score(y_test, y_pred):.4f}")
print(f"Precision: {precision_score(y_test, y_pred):.4f}")
print(f"Recall: {recall_score(y_test, y_pred):.4f}")
print(f"ROC-AUC: {roc_auc_score(y_test, y_proba):.4f}")

## Обучим модель на ребалансированном датасете

from imblearn.over_sampling import SMOTE
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline

over = SMOTE(sampling_strategy=0.1, random_state=42)  # Доводим меньший класс до 10% 
under = RandomUnderSampler(sampling_strategy=0.5)     # Уменьшаем больший класс

pipeline = Pipeline([
    ('o', over),
    ('u', under)
])

X_train_resampled, y_train_resampled = pipeline.fit_resample(X_train, y_train)

rf2 = RandomForestClassifier(
    random_state=42,
    # class_weight='balanced',
    n_jobs=-1
)

param_grid = {
    'n_estimators': [10, 50],
    'max_depth': [5, 10, 20], 
}

cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

grid_search_2 = GridSearchCV(
    estimator=rf,
    param_grid=param_grid,
    scoring='f1',
    cv=cv,
    n_jobs=-1,
    verbose=2
)

grid_search_2.fit(X_train_resampled, y_train_resampled)

y_pred = grid_search_2.predict(X_test)
y_proba = grid_search_2.predict_proba(X_test)[:, 1]

print(f"Подобранные параметры (модель 2): {grid_search_2.best_params_}")
print(f"F1-score (CV): {grid_search_2.best_score_:.4f}")
print("\nМетрики:")
print(f"F1-score: {f1_score(y_test, y_pred):.4f}")
print(f"Precision: {precision_score(y_test, y_pred):.4f}")
print(f"Recall: {recall_score(y_test, y_pred):.4f}")
print(f"ROC-AUC: {roc_auc_score(y_test, y_proba):.4f}")

## Выводы:

Модель `RandomForest` не дает лучших результатов над бэйзлайн моделью логистистической регрессии на данном датасете. Даже метрика `recall` в лучшем случае `RandomForest` равна `0.4149`. `Precision` по-прежнему не удалось хоть сколько-нибудь повысить, эта метрика остается на минимальном уровне (`~0.02`)
